consolidate geom icons into a util
