+++ app/templates/components/project-geometries.hbs
@@ -75,19 +75,19 @@
   }}
     <ul class="over-map-legend no-bullet">
       {{#if (not (is-empty model.developmentSite))}}
-        <li><strong>Development Site</strong> {{labs-ui/legend-icon icon=developmentSiteIcon}}</li>
+        <li><strong>Development Site</strong> {{labs-ui/legend-icon icon=projectGeometryIcons.developmentSiteIcon}}</li>
       {{/if}}
       {{#if (not (is-empty model.projectArea))}}
-        <li><strong>Project Area</strong> {{labs-ui/legend-icon icon=projectAreaIcon}}</li>
+        <li><strong>Project Area</strong> {{labs-ui/legend-icon icon=projectGeometryIcons.projectAreaIcon}}</li>
       {{/if}}
       {{#if (not (is-empty model.underlyingZoning))}}
-        <li><strong>Underlying Zoning</strong> {{labs-ui/legend-icon icon=underlyingZoningIcon}}</li>
+        <li><strong>Underlying Zoning</strong> {{labs-ui/legend-icon icon=projectGeometryIcons.underlyingZoningIcon}}</li>
       {{/if}}
       {{#if (not (is-empty model.specialPurposeDistricts))}}
-        <li><strong>Special Purpose Districts</strong> {{labs-ui/legend-icon icon=specialPurposeDistrictsIcon}}</li>
+        <li><strong>Special Purpose Districts</strong> {{labs-ui/legend-icon icon=projectGeometryIcons.specialPurposeDistrictsIcon}}</li>
       {{/if}}
       {{#if (not (is-empty model.rezoningArea))}}
-        <li><strong>Rezoning Area</strong> {{labs-ui/legend-icon icon=rezoningAreaIcon}}</li>
+        <li><strong>Rezoning Area</strong> {{labs-ui/legend-icon icon=projectGeometryIcons.rezoningAreaIcon}}</li>
       {{/if}}
     </ul>
   {{/if}}
