+++ app/templates/projects/show.hbs
@@ -36,7 +36,7 @@
         <div class="cell auto">
           <strong>Development Site</strong>
           {{#if (not (is-empty model.developmentSite))}}
-            {{labs-ui/legend-icon icon=developmentSiteIcon}}
+            {{labs-ui/legend-icon icon=projectGeometryIcons.developmentSiteIcon}}
           {{/if}}
         </div>
         <div class="cell shrink">
@@ -83,7 +83,7 @@
         <div class="cell auto">
           <strong>Project Area</strong>
           {{#if (not (is-empty model.projectArea))}}
-            {{labs-ui/legend-icon icon=projectAreaIcon}}
+            {{labs-ui/legend-icon icon=projectGeometryIcons.projectAreaIcon}}
           {{/if}}
         </div>
         <div class="cell shrink">
@@ -130,7 +130,7 @@
         <div class="cell auto">
           <strong>Underlying Zoning</strong>
           {{#if (not (is-empty model.underlyingZoning))}}
-            {{labs-ui/legend-icon icon=underlyingZoningIcon}}
+            {{labs-ui/legend-icon icon=projectGeometryIcons.underlyingZoningIcon}}
           {{/if}}
         </div>
         <div class="cell shrink">
@@ -199,7 +199,7 @@
         <div class="cell auto">
           <strong>Special Purpose Districts</strong>
           {{#if (not (is-empty model.specialPurposeDistricts))}}
-            {{labs-ui/legend-icon icon=specialPurposeDistrictsIcon}}
+            {{labs-ui/legend-icon icon=projectGeometryIcons.specialPurposeDistrictsIcon}}
           {{/if}}
         </div>
         <div class="cell shrink">
@@ -232,7 +232,7 @@
         <div class="cell auto">
           <strong>Rezoning Area</strong>
           {{#if (not (is-empty model.rezoningArea))}}
-            {{labs-ui/legend-icon icon=rezoningAreaIcon}}
+            {{labs-ui/legend-icon icon=projectGeometryIcons.rezoningAreaIcon}}
           {{/if}}
         </div>
         <div class="cell shrink">
