+++ tests/unit/utils/project-geom-icons-test.js
@@ -1,0 +1,11 @@
+import projectGeomIcons from 'labs-applicant-maps/utils/project-geom-icons';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | project-geom-icons', function(hooks) {
+
+  // Replace this with your real tests.
+  test('it works', function(assert) {
+    let result = projectGeomIcons();
+    assert.ok(result);
+  });
+});
